refactor: 컨트롤러에 맞게 테스트 코드 분리
